pyproject.toml
[project]
name = "2d-ditgital-human"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "gradio>=6.9.0",
    "gradio-client>=2.3.0",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]
